Module onmosterspawn changed to monster_spawn
