+++ MajruszsDifficulty1-16-3/src/main/java/com/majruszs_difficulty/events/monster_spawn/ApplyingNegativePotionEffectsOnCreeper.java
@@ -1,4 +1,4 @@
-package com.majruszs_difficulty.events.onmonsterspawn;
+package com.majruszs_difficulty.events.monster_spawn;
 
 import com.majruszs_difficulty.MajruszsDifficulty;
 import com.majruszs_difficulty.MajruszsHelper;
