+++ MajruszsDifficulty1-16-3/src/main/java/com/majruszs_difficulty/events/monster_spawn/SkeletonGroup.java
@@ -1,22 +1,19 @@
-package com.majruszs_difficulty.events.onmonsterspawn;
+package com.majruszs_difficulty.events.monster_spawn;
 
 import com.majruszs_difficulty.MajruszsDifficulty;
-import com.majruszs_difficulty.MajruszsHelper;
 import net.minecraft.entity.CreatureEntity;
 import net.minecraft.entity.EntityType;
-import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.monster.MonsterEntity;
-import net.minecraft.entity.monster.ZombieEntity;
-import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.entity.monster.SkeletonEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.world.server.ServerWorld;
 
-public class ZombieGroup extends EnemyGroup {
+public class SkeletonGroup extends EnemyGroup {
+	protected static final double woodenSwordChance = 0.5D;
 	protected static final double stoneSwordChance = 0.25D;
-	protected static final double woodenAxeChance = 0.25D;
 
-	public ZombieGroup( MonsterEntity leader, ServerWorld world ) {
+	public SkeletonGroup( MonsterEntity leader, ServerWorld world ) {
 		super( leader, world, 1, 3 );
 
 		giveArmorToLeader( world, Armors.leather );
@@ -24,20 +21,20 @@ public class ZombieGroup extends EnemyGroup {
 
 	@Override
 	protected CreatureEntity spawnChild( ServerWorld world ) {
-		ZombieEntity zombie = EntityType.ZOMBIE.create( world );
+		SkeletonEntity skeleton = EntityType.SKELETON.create( world );
 
-		giveWeaponTo( zombie, world );
-		setupGoals( zombie, 9, 9 );
+		giveWeaponTo( skeleton, world );
+		setupGoals( skeleton, 9, 9 );
 
-		return zombie;
+		return skeleton;
 	}
 
 	protected ItemStack generateWeapon() {
 		double itemChance = MajruszsDifficulty.RANDOM.nextDouble();
 
-		if( itemChance <= woodenAxeChance )
-			return new ItemStack( Items.WOODEN_AXE );
-		else if( itemChance <= woodenAxeChance + stoneSwordChance )
+		if( itemChance <= woodenSwordChance )
+			return new ItemStack( Items.WOODEN_SWORD );
+		else if( itemChance <= woodenSwordChance + stoneSwordChance )
 			return new ItemStack( Items.STONE_SWORD );
 		else
 			return null;
